Prefer authorized time over posted time
Plaid's guidance is to prefer the authorized timestamp over the
posted timestamp when showing transactions to users. The reasoning
for this is that the authorized timestamp will be closer to the
actual time that the user initiated the transaction. This commit
updates the Firefly "date" field to prefer the authorized time
if it's available. The posted time will be saved to the Firefly
processDate field instead.